requirements.txt
langgraph==0.2.16
langchain==0.2.16
langchain-openai==0.1.25
pandas==2.2.1
numpy==1.26.4
matplotlib==3.8.4
seaborn==0.13.2
scikit-learn==1.7.1
pillow==10.2.0
openai==1.58.1
python-dotenv==1.0.1 